Backend/requirements.txt
flask
flask-sqlalchemy
flask-restful
flask-restx
flask-marshmallow
marshmallow-sqlalchemy
jsonschema
marshmallow
flask-jwt-extended
flask-bcrypt
flask-cors
python-dotenv
faker 
Flask-Migrate
mkdocs
mkdocs-material
mkdocs-swagger-ui
mkdocs-mermaid2-plugin

